--- src/__test_fixtures__/README.md
@@ -1,1 +1,0 @@
-Add fixtures used during testing here
