+++ src/pages/__tests__/Home.test.tsx
@@ -2,7 +2,7 @@
 import { axe } from 'jest-axe';
 import { render } from '@testing-library/react';
 
-import Home from '../../pages/Home';
+import Home from '../Home';
 
 describe('Home component tests', () => {
   test('Home component is accessible', async () => {
